Refactor useCreateCashuWallet into a separate hook file
