+++ src/hooks/useCreateCashuWallet.ts
@@ -1,0 +1,48 @@
+import { useMutation } from '@tanstack/react-query';
+import { useCurrentUser } from '@/hooks/useCurrentUser';
+import { useCashuWallet } from '@/hooks/useCashuWallet';
+import { useCashuStore } from '@/stores/cashuStore';
+import { derivePrivkeyFromNostrSignature } from '@/lib/nip60';
+import { defaultMints } from '@/lib/cashu';
+
+/**
+ * Hook for creating a Cashu wallet using the user's Nostr identity
+ * 
+ * @returns A mutation for creating a Cashu wallet
+ */
+export function useCreateCashuWallet() {
+  const { user } = useCurrentUser();
+  const { createWallet } = useCashuWallet();
+  const cashuStore = useCashuStore();
+
+  return useMutation({
+    mutationFn: async () => {
+      if (!user) {
+        throw new Error('You must be logged in to create a wallet');
+      }
+
+      try {
+        const privkey = await derivePrivkeyFromNostrSignature(
+          user.signer,
+          user.pubkey
+        );
+        cashuStore.setPrivkey(privkey);
+
+        // Create a new wallet with the default mint
+        const mints = cashuStore.mints.map((m) => m.url);
+        // add default mints
+        mints.push(...defaultMints);
+
+        createWallet({
+          privkey,
+          mints,
+        });
+
+        return { success: true };
+      } catch (error) {
+        console.error('Failed to derive private key:', error);
+        throw new Error('Failed to create wallet. Please try again.');
+      }
+    }
+  });
+}
